LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/iframe-document-preserve.window-expected.txt


PASS moveBefore(): about:blank iframe's document is preserved
PASS moveBefore(): simple same-origin document is preserved
PASS moveBefore(): cross-origin iframe is preserved: remove new parent
PASS moveBefore(): cross-origin iframe is preserved: remove self
PASS moveBefore(): cross-origin iframe is preserved: remove self via replaceChildren()
PASS moveBefore(): cross-origin iframe is preserved: remove self via innerHTML
FAIL window.frames ordering does not change due to moveBefore() assert_equals: iframe3 comes first in DOM after moveBefore expected "iframe3" but got "iframe1"
